feat: render html folds also
