+++ src/components/Renderer/index.jsx
@@ -3,7 +3,13 @@ import { useLocation } from "react-router-dom";
 import { gql, useLazyQuery, useQuery, useSubscription } from "@apollo/client";
 import { MenuContext, SettingsContext, CustomerContext } from "../../context";
 import { DailyKit, fileAgent, removeChildren } from "../../utils";
-import { ORDERS, ALL_COUPONS, CAMPAIGNS, PRODUCTS, CARTS } from "../../graphql";
+import {
+  SUBSCRIPTION,
+  ALL_COUPONS,
+  CAMPAIGNS,
+  PRODUCTS,
+  CARTS,
+} from "../../graphql";
 import { Loader } from "..";
 
 const Renderer = ({ moduleId, moduleType, moduleConfig, moduleFile }) => {
@@ -38,12 +44,12 @@ const Renderer = ({ moduleId, moduleType, moduleConfig, moduleFile }) => {
     }
   );
 
-  const { loading: ordersQueryLoading, error, data } = useSubscription(
-    gql(ORDERS),
+  const { loading: ordersQueryLoading, error } = useSubscription(
+    gql(SUBSCRIPTION.ORDERS.FETCH),
     {
       variables: {
-        brandId: 1,
-        keycloakId: "33da8306-e5eb-4cb5-bae9-9327fd7700d6",
+        brandId: settings?.brand?.id,
+        keycloakId: customer?.customer?.keycloakId,
       },
       onSubscriptionData: ({
         subscriptionData: { data: { orders = [] } = {} } = {},
@@ -57,7 +63,7 @@ const Renderer = ({ moduleId, moduleType, moduleConfig, moduleFile }) => {
 
   const { loading: couponLoading } = useQuery(gql(ALL_COUPONS), {
     variables: {
-      brandId: 1,
+      brandId: settings?.brand?.id,
     },
     onCompleted: ({ coupons = [] }) => {
       console.log(coupons);
@@ -70,7 +76,7 @@ const Renderer = ({ moduleId, moduleType, moduleConfig, moduleFile }) => {
 
   const { loading: campaignLoading } = useQuery(gql(CAMPAIGNS), {
     variables: {
-      brandId: 1,
+      brandId: settings?.brand?.id,
     },
     onCompleted: ({ campaigns = [] }) => {
       console.log(campaigns);
@@ -160,7 +166,7 @@ const Renderer = ({ moduleId, moduleType, moduleConfig, moduleFile }) => {
       } catch (error) {
         console.log(error);
       }
-      // console.log("checking...", hydratedMenu, menu);
+      console.log("checking...", hydratedMenu, settings);
       const parsedHtml = await DailyKit.engine(moduleFile.path, {
         ...{ cart: customer.cart },
         ...settings,
