+++ src/graphql/queries/index.js
@@ -125,35 +125,6 @@ query Customer($keycloakId: String!) {
 }
 `;
 
-export const ORDERS = `
-   subscription Orders($keycloakId: String!, $brandId: Int!) {
-      orders(
-         where: {
-            keycloakId: { _eq: $keycloakId }
-            brandId: { _eq: $brandId }
-         }
-         order_by: { created_at: desc }
-      ) {
-         id
-         deliveryInfo
-         fulfillmentType
-         orderStatus
-         amountPaid
-         currency
-         deliveryPrice
-         transactionId
-         discount
-         tax
-         tip
-         itemTotal
-         created_at
-         orderCart {
-            cartInfo
-         }
-      }
-   }
-`;
-
 export const ALL_COUPONS = `
   query Coupons($brandId: Int!) {
     coupons(
